+++ parser_library/src/diagnostic.cpp
@@ -2629,44 +2629,54 @@ diagnostic_s diagnostic_s::error_B4G002(const utils::resource::resource_location
         diagnostic_tag::none);
 }
 
-diagnostic_s diagnostic_s::error_L0001(std::string_view path)
-{
-    return diagnostic_s("", {}, "L0001", concat("Unable to load library: ", path, "."));
-}
-
-diagnostic_s diagnostic_s::error_L0002(std::string_view path)
+diagnostic_s diagnostic_s::error_L0001(
+    const utils::resource::resource_location& config_loc, const utils::resource::resource_location& lib_loc)
 {
     return diagnostic_s(
-        "", {}, "L0002", concat("Unable to load library: ", path, ". Error: The path does not point to directory."));
+        config_loc.get_uri(), {}, "L0001", concat("Unable to load library: ", lib_loc.to_presentable(), "."));
 }
 
-diagnostic_s diagnostic_s::warning_L0003(std::string_view path)
+diagnostic_s diagnostic_s::error_L0002(
+    const utils::resource::resource_location& config_loc, const utils::resource::resource_location& lib_loc)
 {
-    return diagnostic_s("",
+    return diagnostic_s(config_loc.get_uri(),
+        {},
+        "L0002",
+        concat("Unable to load library: ", lib_loc.to_presentable(), ". Error: The path does not point to directory."));
+}
+
+diagnostic_s diagnostic_s::warning_L0003(
+    const utils::resource::resource_location& config_loc, const utils::resource::resource_location& lib_loc)
+{
+    return diagnostic_s(config_loc.get_uri(),
         {},
         diagnostic_severity::warning,
         "L0003",
         concat("Macros from library '",
-            path,
+            lib_loc.to_presentable(),
             "' were selected by a deprecated mechanism to specify file extensions (alwaysRecognize in pgm_conf.json)."),
         {},
         diagnostic_tag::none);
 }
 
-diagnostic_s diagnostic_s::warning_L0004(std::string_view path, std::string_view macro_name)
+diagnostic_s diagnostic_s::warning_L0004(const utils::resource::resource_location& config_loc,
+    const utils::resource::resource_location& lib_loc,
+    std::string_view macro_name)
 {
-    return diagnostic_s("",
+    return diagnostic_s(config_loc.get_uri(),
         {},
         diagnostic_severity::warning,
         "L0004",
-        concat("Library '", path, "' contains multiple definitions of the macro '", macro_name, "'."),
+        concat(
+            "Library '", lib_loc.to_presentable(), "' contains multiple definitions of the macro '", macro_name, "'."),
         {},
         diagnostic_tag::none);
 }
 
-diagnostic_s diagnostic_s::warning_L0005(std::string_view pattern, size_t limit)
+diagnostic_s diagnostic_s::warning_L0005(
+    const utils::resource::resource_location& config_loc, std::string_view pattern, size_t limit)
 {
-    return diagnostic_s("",
+    return diagnostic_s(config_loc.get_uri(),
         {},
         diagnostic_severity::warning,
         "L0005",
@@ -2675,9 +2685,9 @@ diagnostic_s diagnostic_s::warning_L0005(std::string_view pattern, size_t limit)
         diagnostic_tag::none);
 }
 
-diagnostic_s diagnostic_s::warning_L0006(std::string_view path)
+diagnostic_s diagnostic_s::warning_L0006(const utils::resource::resource_location& config_loc, std::string_view path)
 {
-    return diagnostic_s("",
+    return diagnostic_s(config_loc.get_uri(),
         {},
         diagnostic_severity::warning,
         "L0006",