+++ parser_library/src/workspaces/library_local.h
@@ -41,8 +41,10 @@ class library_local final : public library, public diagnosable_impl
 public:
     // takes reference to file manager that provides access to the files
     // and normalised path to directory that it wraps.
-    library_local(
-        file_manager& file_manager, utils::resource::resource_location lib_loc, library_local_options options);
+    library_local(file_manager& file_manager,
+        utils::resource::resource_location lib_loc,
+        library_local_options options,
+        utils::resource::resource_location proc_grps_loc);
 
     library_local(const library_local&) = delete;
     library_local& operator=(const library_local&) = delete;
@@ -60,16 +62,17 @@ public:
     bool is_once_only() const override { return false; }
 
 private:
-    file_manager& file_manager_;
+    file_manager& m_file_manager;
 
-    utils::resource::resource_location lib_loc_;
+    utils::resource::resource_location m_lib_loc;
     std::unordered_map<std::string, utils::resource::resource_location, utils::hashers::string_hasher, std::equal_to<>>
-        files_;
-    std::vector<std::string> extensions_;
+        m_files;
+    std::vector<std::string> m_extensions;
     // indicates whether load_files function was called (not whether it was successful)
-    bool files_loaded_ = false;
-    bool optional_ = false;
-    bool extensions_from_deprecated_source = false;
+    bool m_files_loaded = false;
+    bool m_optional = false;
+    bool m_extensions_from_deprecated_source = false;
+    utils::resource::resource_location m_proc_grps_loc;
 
     void load_files();
 };